create custom hook to fetch data
